requirements.txt
Django==3.1.5
django-environ==0.4.5
django-widget-tweaks==1.4.8
mysqlclient==2.0.3
django-allauth==0.44.0
psycopg2-binary==2.8.6
celery==5.2.7
pyodbc==4.0.30
django-tenants==3.4.4
Pillow==9.2.0
django-cities-light==3.9.2
django-celery-beat==2.3.0
#For SSO PART Please manually install: 
# django-auth-adfs
# redis==4.3.4
#Production
gunicorn==20.0.4
whitenoise
django-storages[azure]
importlib-metadata==4.13.0
pandas
azure-data-tables
Redis
six
openpyxl